requirement.txt
beautifulsoup4==4.12.3
bs4==0.0.2
fastapi==0.111.0
langchain==0.2.0
pymilvus==2.4.3
python-decouple==3.8
requests==2.32.2
uvicorn==0.29.0
langchain-text-splitters
retrying
openai
